# LetsCode API.OpenAPI3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LetsCode API",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    },
    {
      "url": "https://lets-code-site.herokuapp.com",
      "variables": {}
    }
  ],
  "paths": {
    "/player": {
      "post": {
        "tags": [
          "localhost"
        ],
        "summary": "Create player",
        "operationId": "Createplayer",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateplayerRequest"
              },
              "example": {
                "name": "new player 2",
                "login": "postman_test2",
                "password": "postman123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      },
      "get": {
        "tags": [
          "localhost"
        ],
        "summary": "ListAll players",
        "operationId": "ListAllplayers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
    "/player/login": {
      "post": {
        "tags": [
          "localhost"
        ],
        "summary": "Login for start game",
        "operationId": "Loginforstartgame",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginforstartgameRequest"
              },
              "example": {
                "name": "new player2",
                "login": "postman_test2",
                "password": "postman123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:8080",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateplayerRequest": {
        "title": "CreateplayerRequest",
        "required": [
          "name",
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "name": "new player 2",
          "login": "postman_test2",
          "password": "postman123"
        }
      },
...
      "LoginforstartgameRequest": {
        "title": "LoginforstartgameRequest",
        "required": [
          "name",
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "name": "new player2",
          "login": "postman_test2",
          "password": "postman123"
        }
      },
...
    }
  },
...
}